+++ projects/qgrid-core/row-details/row.details.status.d.ts
@@ -1,5 +1,5 @@
 export declare class RowDetailsStatus {
 	constructor(expand: boolean);
 
-	expand: boolean;
+	readonly expand: boolean;
 }