+++ src/examples/details-row-select/example-details-row-select.component.ts
@@ -1,0 +1,20 @@
+import { Component, ChangeDetectionStrategy } from '@angular/core';
+import { DataService, Atom } from '../data.service';
+import { Observable } from 'rxjs';
+
+@Component({
+	selector: 'example-details-row-select',
+	templateUrl: 'example-details-row-select.component.html',
+	styleUrls: ['example-details-row-select.component.scss'],
+	providers: [DataService],
+	changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class ExampleDetailsRowSelectComponent {
+	static id = 'details-row-select';
+
+	rows: Observable<Atom[]>;
+
+	constructor(dataService: DataService) {
+		this.rows = dataService.getAtoms();
+	}
+}